risolto commenti pull request, io lascerei l'uso di pydantic per l'implementazione backend
